src/test/java/ru/job4j/io/ConfigTest.java
package ru.job4j.io;

import org.junit.jupiter.api.Test;
import static org.assertj.core.api.Assertions.*;

class ConfigTest {

    @Test
    void whenPairWithoutComment() {
        String path = "./data/pair_without_comment.properties";
        Config config = new Config(path);
        config.load();
        assertThat(config.value("name")).isEqualTo("Liana");
    }

    @Test
    void whenPairWithComment() {
        String path = "./data/pair_with_comment.properties";
        Config config = new Config(path);
        config.load();
        assertThat(config.value("name")).isEqualTo("Liana");
        assertThat(config.value("age")).isEqualTo("76");
    }

    @Test
    void whenPairWithEmptyLine() {
        String path = "./data/pair_with_empty_lines.properties";
        Config config = new Config(path);
        config.load();
        assertThat(config.value("name")).isEqualTo("Liana");
        assertThat(config.value("age")).isEqualTo("90");
    }

    @Test
    void whenPairInvalid() {
        String path = "./data/pair_with_start_with=_lines.properties";
        Config config = new Config(path);
        assertThatThrownBy(config::load)
                .isInstanceOf(IllegalArgumentException.class)
                .hasMessageContaining("pattern violation");
    }
}
